replace with team format
