--- packages/jaree-server-repo/src/main/java/org/jaree/api/application/inputs/ApplicationContextSaveInputDTO.java
@@ -1,18 +1,0 @@
-package org.jaree.api.application.inputs;
-
-import java.time.LocalDateTime;
-
-import lombok.Data;
-
-@Data
-public class ApplicationContextSaveInputDTO {
-  private Long id;
-  private LocalDateTime createdAt;
-  private String s3Link;
-  private Question question;
-
-  @Data
-  public static class Question {
-    private String id;
-  }
-}
